Split test_embargo.py (received) into focused files by concern
Split the 1197-line test/core/use_cases/received/test_embargo.py (4
classes, 24 tests) into four smaller files grouped by use-case type:

- test_embargo_propose_accept_reject.py  (9 tests): create/invite,
  accept, and reject lifecycle (TestEmbargoProposalLifecycle)
- test_embargo_term_revise.py           (5 tests): add/remove embargo
  and unusual-state transitions (TestEmbargoTermRevise)
- test_embargo_misc.py                  (4 tests): Announce no-op
  receiver and #609 consent-reset regression tests
  (TestAnnounceEmbargoEventToCaseReceivedUseCase,
   TestResetEmbargoConsentWithInlineParticipants)
- test_embargo_ledger_cascade.py        (6 tests): CaseLedgerEntry
  cascade for all embargo received-side handlers (TestEmbargoLogEntryCascade)

Pure reorganization — no logic changes. All 24 tests pass unchanged.

Closes #1089
